Add dedicated conversation flow module structure
