+++ public/flows/builder.html
@@ -1,0 +1,1 @@
+<h1>Builder de Fluxo</h1>
