+++ public/flows/builder.js
@@ -1,0 +1,1 @@
+// JS para builder
