+++ public/flows/flows.css
@@ -1,0 +1,1 @@
+/* Estilos para fluxos */
