+++ public/flows/flows.js
@@ -1,0 +1,1 @@
+// JS para lista de fluxos
