+++ src/app.js
@@ -7,7 +7,7 @@ const pedidosController = require('./controllers/pedidosController');
 const envioController = require('./controllers/envioController');
 const rastreamentoController = require('./controllers/rastreamentoController');
 const automationsController = require('./controllers/automationsController');
-const flowsController = require('./controllers/flowsController');
+const flowsController = require('./flows/flowController');
 const integrationsController = require('./controllers/integrationsController');
 const logsController = require('./controllers/logsController');
 const paymentController = require('./controllers/paymentController');
